% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.utsf}
\alias{autoplot.utsf}
\title{Create a ggplot object from an \code{utsf} object}
\usage{
\method{autoplot}{utsf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{utsf}.}

\item{...}{additional parameter.}
}
\value{
The \code{ggplot} object representing a plotting of the time series and
its forecast.
}
\description{
Plot the time series and its associated forecast.
}
\examples{
f <- forecast(AirPassengers, h = 12, lags = 1:12, method = "rf")
library(ggplot2)
autoplot(f)
}
