% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.utsf_forecast}
\alias{autoplot.utsf_forecast}
\title{Create a ggplot object from an \code{utsf_forecast} object}
\usage{
\method{autoplot}{utsf_forecast}(object, ...)
}
\arguments{
\item{object}{An object of class \code{utsf_forecast}.}

\item{...}{additional parameter.}
}
\value{
The \code{ggplot} object representing a plotting of the time series and
its forecast.
}
\description{
Plot the time series and its associated forecast.
}
\examples{
m <- create_model(AirPassengers, lags = 1:12, method = "rf")
f <- forecast(m, h = 12)
library(ggplot2)
autoplot(f)
}
