% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_differences.R
\name{differences}
\alias{differences}
\title{Specifying the order of differences}
\usage{
differences(n = -1)
}
\arguments{
\item{n}{An integer specifying the order of differences to be applied.
If the default (-1) is used, the order of differences needed by the
time series will be computed by the \code{\link[forecast:ndiffs]{forecast::ndiffs()}} function.}
}
\value{
An integer with the order of differences to be applied.
}
\description{
This function is used to specify that the time series will be preprocessed
using first differences.
}
\examples{
differences(1)
}
