% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_accuracy.R
\name{evaluation}
\alias{evaluation}
\title{Specify the way in which the estimation of forecast accuracy is done}
\usage{
evaluation(type = "normal", size = NULL, prop = NULL)
}
\arguments{
\item{type}{A string. Possible values are \code{"normal"} (the default) and
\code{"minimum"}. See the vignette for an explanation of both ways of evaluating
forecast accuracy.}

\item{size}{An integer. It is the size of the test set (how many of the last
observations of the time series are used as test set). It can only be used
when the type parameter is \code{"normal"}.}

\item{prop}{A numeric value in the range (0, 1). It is the proportion of the
time series used as test set. It can only be used when the type parameter
is \code{"normal"}.}
}
\value{
An S3 object of class \code{evaluation} with the chosen evaluation.
}
\description{
Specify the way in which the estimation of forecast accuracy is done
}
\examples{
evaluation("normal", size = 10)
}
