% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_differences.R
\name{fd}
\alias{fd}
\title{Specifying the number of first differences}
\usage{
fd(n = -1)
}
\arguments{
\item{n}{An integer specifying the number of first differences to be applied.
If the default (-1) is used, the number of first differences needed by the
time series will be computed by the \code{\link[forecast:ndiffs]{forecast::ndiffs()}} function.}
}
\value{
An integer with the number of first differences to be applied.
}
\description{
This function is used to specify that the time series will be preprocessed
using first differences.
}
\examples{
fd(1)
}
