% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Fit an univariate time series forecasting model and make forecasts}
\usage{
forecast(
  timeS,
  h,
  lags = NULL,
  method = "knn",
  param = NULL,
  transform = "additive"
)
}
\arguments{
\item{timeS}{A time series of class \code{ts} or a numeric vector.}

\item{h}{A positive integer. Number of values to be forecast into the future,
i.e., forecast horizon.}

\item{lags}{An integer vector, in increasing order, expressing the lags used
as autoregressive variables. If the default value (\code{NULL}) is provided, a
suitable vector is chosen.}

\item{method}{A string indicating the method used for training and
forecasting. Allowed values are:
\itemize{
\item \code{"knn"}: k-nearest neighbors (the default)
\item \code{"rt"}: regression trees
\item \code{"mt"}:  model trees
\item \code{"bagging"}
\item \code{"rf"}: random forest.
}

See details for a brief explanation of the models.}

\item{param}{A list with parameters for the underlying function that builds
the model. If the default value (\code{NULL}) is provided, the model is built
with its default parameters. See details for the functions used to train
the models.}

\item{transform}{A character value indicating whether the training samples
are transformed. If the time series has a trend it is recommended. By
default is \code{"additive"} (additive transformation). It is also possible a
multiplicative transformation or no transformation.}
}
\value{
A S3 object of class \code{utsf}, basically a list with, at least, the
following components: \item{\code{ts}}{The time series being forecast.}
\item{\code{features}}{A data frame with the features of the training set. The
column names of the data frame indicate the autoregressive lags.}
\item{\code{targets}}{A vector with the targets of the training set.}
\item{\code{lags}}{An integer vector with the autoregressive lags.}
\item{\code{model}}{The regression model used recursively to make the forecast.}
\item{\code{pred}}{An object of class \code{ts} and length \code{h} with the forecast.}
}
\description{
This function trains a model from the historical values of a time series
using as targets the historical values and as features of targets their
lagged values. Then, the fitted model is used to predict the future values of
the series using a recursive strategy.
}
\details{
The functions used to build and train the model are:
\itemize{
\item KNN: In this case no model is built and the function \code{\link[FNN:knn.reg]{FNN::knn.reg()}} is
used to predict the future values of the time series.
\item Regression trees: Function \code{\link[rpart:rpart]{rpart::rpart()}} to build the model and the
method \code{\link[=predict.rpart]{predict.rpart()}} associated with the trained model to forecast the
future values of the time series.
\item Model trees: Function \code{\link[Cubist:cubist.default]{Cubist::cubist()}} to build the model and the
method \code{\link[=predict.cubist]{predict.cubist()}} associated with the trained model to forecast the
future values of the time series.
\item Bagging: Function \code{\link[ipred:bagging]{ipred::bagging()}} to build the model and the
method \code{\link[=predict.regbagg]{predict.regbagg()}} associated with the trained model to forecast the
future values of the time series.
\item Random forest: Function \code{\link[ranger:ranger]{ranger::ranger()}} to build the model and the
method \code{\link[=predict.ranger]{predict.ranger()}} associated with the trained model to forecast the
future values of the time series.
}
}
\examples{
## Forecast time series using k-nearest neighbors
forecast(AirPassengers, h = 12, method = "knn")$pred

## Forecast time series using k-nearest neighbors changing the default k
forecast(AirPassengers, h = 12, method = "knn", par = list(k = 5))$pred
}
