% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_models.R
\name{predict.utsf}
\alias{predict.utsf}
\title{Prediction from \code{utsf} objects}
\usage{
\method{predict}{utsf}(object, new_value, ...)
}
\arguments{
\item{object}{object of class \code{utsf}.}

\item{new_value}{a data frame with one row of a new observation.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a numeric value with the forecast.
}
\description{
Predict the class of a new observation based on the model associated with the
\code{utsf} object
}
