% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utsf-package.R
\docType{package}
\name{utsf-package}
\alias{utsf}
\alias{utsf-package}
\title{utsf: Univariate Time Series Forecasting}
\description{
An engine for univariate time series forecasting using different regression models in an autoregressive way. The engine provides an uniform interface for applying the different models. Furthermore, it is extensible so that users can easily apply their own regression models to univariate time series forecasting and benefit from all the features of the engine, such as preprocessings or estimation of forecast accuracy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franciscomartinezdelrio/utsf}
  \item Report bugs at \url{https://github.com/franciscomartinezdelrio/utsf/issues}
}

}
\author{
\strong{Maintainer}: Francisco Martinez \email{fmartin@ujaen.es} (\href{https://orcid.org/0000-0002-5206-1898}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Maria Pilar Frias-Bustamante \email{mpfrias@ujaen.es} (\href{https://orcid.org/0000-0001-6886-0953}{ORCID})
}

}
\keyword{internal}
