% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Fit an Univariate Time Series Forcasting Model and Forecast}
\usage{
forecast(timeS, h, lags = NULL, method, param = NULL, transform = "additive")
}
\arguments{
\item{timeS}{A numeric vector or time series of class \code{ts}.}

\item{h}{A positive integer. Number of values to forecast.}

\item{lags}{An integer vector in increasing order expressing the lags used as
autoregressive variables.}

\item{method}{A string indicating the method used for forecasting. Allowed
values are "knn", "rt" (regression trees), "mt" (model trees), "bagging"
and "rf" (random forest).}

\item{param}{A list with parameters for the underlying function that creates
the model.}

\item{transform}{A character value indicating whether the training samples
are transformed. If the time series has a trend it is recommended. By
default is \code{"additive"} (additive transformation). It is also possible
a multiplicative transformation or no transformation.}
}
\value{
A list with values
}
\description{
To be done
}
\examples{
## Forecast time series using k-nearest neighbors
forecast(AirPassengers, h = 12, method = "knn")$pred
}
