% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FORECAST}
\alias{FORECAST}
\title{Forecasting time series}
\usage{
FORECAST(model, h)
}
\arguments{
\item{model}{an object of class \code{utsf} embedding a forecasting model for a
time series.}

\item{h}{A positive integer. Number of values to be forecast into the future,
i.e., forecast horizon.}
}
\value{
an object of class \code{utsf_forecast} with the same components of the
model received as first argument, plus a component named \code{pred} containing
the forecast as an \code{ts} object.
}
\description{
Forecasting time series
}
\examples{
## Forecast time series using k-nearest neighbors
m <- create_model(AirPassengers, method = "knn")
f <- FORECAST(m, h = 12)
f$pred
library(ggplot2)
autoplot(f)

## Using k-nearest neighbors changing the default k value
m <- create_model(AirPassengers, method = "knn", param = list(k = 5))
FORECAST(m, h = 12)

## Using your own regression model

# Function to build the regression model
my_knn_model <- function(X, y) {
  structure(list(X = X, y = y), class = "my_knn")
}
# Function to predict a new example
predict.my_knn <- function(object, new_value) {
  FNN::knn.reg(train = object$X, test = new_value, y = object$y)$pred
}
m <- create_model(AirPassengers, method = my_knn_model)
FORECAST(m, h = 12)
}
